#js - edit createSongElement & createElement
1) createSongElement -
I added to the "children" array the values ​​of the "keys" (title, artist, album, duration and cover) of a defined song.
For each song I created a class - "song" & "box" for future work.

2)createElement -
The function gets an array of "kids" of "classes" and "attributes" and from them creates a new element.
For the array of children - it creates a "text node" for each one that contains the transmitted information.
For the array of classes it defines a class for the created element. For each "attribute" it sets it to the newly created element.

3) durationToMMSS -
The function takes a song duration in seconds and converts it to a string in the format "mm: ss"