#js- edit createPlaylistElement by the requirements + 3 assist functions
1)createPlaylistElement -
The children are (name, number of songs and playlistduration).

2) playlistDuration -
The function receives a playlist id and returns the total duration of the playlist in "mm: ss" format

3)getSongByID -
The function receives a song id and returns a matching song object.

4)getPlaylistLocationByID -
The function receives a playlist id and returns its place in the player.playlists array