# js - playSong() + #CSS - song styles
1) I added id attribute for the songs in the createSongElement. So we can go to song with unique ID on the playSong() function.

2)createElement -
I changed the way to insert the "attributes" into the element so that it would insert Id and "onclick" separately each and its value.

3)playSong -
First the function checks if there is already one element with the class "playing" then it will cancel it. Then it will assign to the song that is clicked the class playing and it will get the style defined in css for it. So only one song will contain the properties of playing at a time.