#JS - bonus 1 + assistanc function
1) playingSong -
When the function is activated, another function (next song) will be activated at a time difference (duration of the current song in seconds) which takes care of transferring to the next song.

2)nextSong-
The function checks whether the song transferred to it has the next song, if not it will display a message that the songs are over. If so she will add the class to the song, set this song as a playable song and run a timeout again for the current song.