#JS - removeSong + 2 assiste functions
1) removeSong-
The function gets id and removes the song from player.songs array and inside the DOM in the song location.
It also removes it from the relevant playlists and updates the changes in the DOM playlist area.

2)getSongLocationByID-
The function gets a song Id and returns the index in the player's song array

3)removeSongFromPlaylist-
The function gets index and id and removes the song from the relevant playlist if it exists in it.